firmware.Config: Ignore hyphenated suffix
Lots of devices have a name with a hyphen-delimited suffix after the
actual board name. Those devices are unable to find their Config file,
and so fail firmware.Config. Here are a few examples from go/tastboard:
* atlas-blueznext
* atlas-kvm
* betty-pi-arc
* caroline-kernelnext
* caroline-ndktranslation
...and more, most of which are variations on those themes.

With this change, we ignore anything after the first hyphen before
loading the config file.

BUG=b:155812495
TEST=firmware.Config
TEST=firmware.Config, but spoof the platform to be "samus-kernelnext"

Change-Id: I151934b874473baebb51b959f5ba59ef89d58b0e