Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> +Sean: Check out the builders that cq-orchestrator marked as relevant for this tast change (context: crbug.com/1077828#c3). It seems pretty random to me.

This is WAI, and is based on rules we've had in place for many months. The builders selected here are the ones for which Tast HW/VM testing is currently configured in CQ. We don't bother doing builds for boards which wouldn't be Tast-tested in CQ anyway, since zero value would be added.

http://cs/chromeos_internal/infra/config/builderconfig/general_config.star?l=121-127&rcl=aa7242026bb7266613ace9f7de0736fcb68c0f90

http://cs/chromeos_internal/infra/config/testingconfig/target_test_requirements_config.star

Patch-set: 1