Set the timeout to chrome.New
Sometimes, a test may fail on chrome.New on some unexpected UI in
the login screen. In that case, currently the test may be stuck
until the given context expires; which is usually the full timeout
of a test case. And sometimes, test timeout is far longer, like
10 minutes or 20 minutes. This causes to run out the time of
entire test runner and causes "Got global error" on further
test cases.

Either way it is not an ideal situation. This CL sets the timeout
to the chrome.New function itself, so if the test can stop earlier
when it gets stuck in the login flow.

This CL checks the loginMode, and sets the maximum timeout
to perform the initialization of chrome; normally 1 minutes but
a bit longer for gaia-login, assuming it may take a bit more
time.

BUG=chromium:1078873
TEST=tast run arc.AuthPerf arc.Boot arc.CachePerf ui.TaskSwitchCUJ

Change-Id: I2b37f832a67327c7aad51b65d1490b023d5f8a08