Requires chrome.New to finish within 3 minutes
Sometimes, a test may fail on chrome.New on some unexpected UI in
the login screen. In that case, currently the test may be stuck
until the given context expires; which is usually the full timeout
of a test case. And sometimes, test timeout is far longer, like
10 minutes or 20 minutes. This causes to run out the time of
entire test runner and causes "Got global error" on further
test cases.

Either way it is not an ideal situation. This CL sets the timeout
to the chrome.New function itself, so if the test can stop earlier
when it gets stuck in the login flow.

I just set the timeout statically, as LoginTimeout * 3 and I
believe it would be safe enough, yet it can save time on such
situation.

BUG=chromium:1078873
TEST=tast run arc.Boot arc.PlayStore'*' ui.TaskSwitchCUJ ui.OOBESmoke

Change-Id: I2b37f832a67327c7aad51b65d1490b023d5f8a08