network.HostapHwsim: add some docs
We only run a limited subset of the network.HostapHwsim tests regularly,
and WiFi developers might run more manually when making large changes to
wpa_supplicant, or when developing new tests for the upstream project.
Document end-to-end how to do this, since it's not a typical workflow
for many.

BUG=none
TEST=view markdown in Gitiles

Change-Id: I10ac58a271e6ffad26d4ad414ed38d7d76acba93
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/2186071
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>
Reviewed-by: Matthew Wang <matthewmwang@chromium.org>
Commit-Queue: Brian Norris <briannorris@chromium.org>
Tested-by: Brian Norris <briannorris@chromium.org>