Make crashpad version of ui.ChromeCrashReporterMetrics.miss
The current version of ui.ChromeCrashReporterMetrics.miss only works we
Breakpad. As we are moving away from Breakpad to Crashpad, we need a
version that works with Crashpad to maintain testing.

We don't have any good levers to stop Crashpad from calling
crash_reporter, so we just kill Crashpad instead to generate a missed
crash report.

BUG=chromium:1036582
Fixed: chromium:1036582
TEST=tast run ... ui.ChromeCrashReporterMetrics.*

Change-Id: I6ed3c6c0109f9cc0d0aa0bc2765ce2b9765ac478