camera: Stabilize camera.CCAUILauncher
This CL does two things:
1. Search "Camera" instead of "camera" in CCAUILauncher
2. Only test minimize/restore behaviors under clamshell mode

Since all apps will automatically minimize when opening launcher under
tablet mode, there is no need to test the minimize/restore behaviors.

However, we still test if the camera can be found in the launcher in
case camera icon is missing in the launcher for tablet mode.

BUG=chromium:1078662
TEST=tast run [DUT] camera.CCAUILauncher (on soraka)
TEST=tast run [DUT] camera.CCAUILauncher (on dumo)

Change-Id: If29f788d0c3e5831fb427be3b563ed0c7ce90fed