arc: stabilize the result of arc.VideoDecodeAccelPerf.
The performance results of arc.VideoDecodeAccelPerf are not quite
stable. The outlier results appears often, and it makes our monitor
service hard to detect the real regression.

In this CL, we run the performance test multiple times, and then
report the median of each metric. It should avoid the outlier result
and make the monitor service more less false alarm.

BUG=b:155944233
TEST=tast run $(DUT) arc.VideoDecodeAccelPerf.*

Change-Id: Icc533dcc3f2328b4c50ec86b17ce1b002758c976