Update patch set 15
Patch Set 15:

> Patch Set 15:
> 
> Hi, 
> 
> I am reviewing the Tastboard results for this test - https://tastboard.corp.google.com/test/security.BootLockbox 
> 
> It is failing frequently. Inspecting the recent failures, one of the problem seems to be the problem of Chrome getting crashed in watcher.go. Any leads on thinking process to solve this ?
Do you see the actual crash log or just the tast message?  If it's Chrome crash, then it's probably not related to this change.  You can search for the stack trace pattern for existing bugs, or file a new one.

In any case, it's easier to discuss in email instead of code review.  Please start another thread if needed.  Thanks!

Patch-set: 15