tast-tests/webrtc: add GPU frequency collection to Perf tests
crrev.com/c/2153720 and introduced GPU command streams performance
collection to graphics; I've been told in the meantime that the
GPU frequency is an important parameter too, so this CL adds
that as well. Command streamer usage and average frequency over
a given workload gives a very interesting view of how the GPU is
behaving.

The GPU frequency collection is somehow strangely logged as
amount-of-cycles (int64) over the collection interval, which is
different from the command-streamer-use-time (time.Duration), so
this CL returns it as a different parameter in dri.go

BUG=chromium:1071841
TEST=tast run DUT webrtc.*Perf.* on kohaku

Change-Id: I05452c02aa8e642fdc87be37af5829aa320e05e6