tast-tests: Fix timing issue in TestTimeline
This test checks that a certain number of samples are taken in a given
time interval. It seems that the sampling interval is too small, and
sometimes we are off by one sample, causing the test to be flaky. This
CL increases the sampling interval.

BUG=chromium:1078850
TEST=/mnt/host/source/src/platform/tast/fast_build.sh -T

Change-Id: Ia3ffae4c0dea07fe386e994a38c64910d89d6658
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/2190053
Reviewed-by: Oleh Lamzin <lamzin@google.com>
Commit-Queue: Oleh Lamzin <lamzin@google.com>
Tested-by: Oleh Lamzin <lamzin@google.com>