Camera test app: Support maximizing window
The camera test app crashes when the user presses the F4 key to maximize
the window. This CL fixes that issue.

There are three changes: First, the fragment should have a constructor
that is not taking any arguments. Second, there are cases in which a
camera exception should be dropped. Third, to be safe, camera tests
should fail if no frames were received by the test app.

BUG=b:156056087
TEST=manual

Change-Id: I2e8fca64a21485732fcba90757f4336268383479