tast-tests: remove spurious error message
Behavior of virtual-usb-printer was changed to not create a record file
unless data to record was sent.

To avoid printing an error message that Remove failed whenever the test
terminates without an existing record file, check that the error
is not an IsNotExist error.

BUG=None
TEST=Run PrintIPPUSB

Change-Id: I9e5acbed283ca085cdfb6bd6fabab62f73f52b23
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/2190853
Tested-by: Fletcher Woodruff <fletcherw@chromium.org>
Commit-Queue: Fletcher Woodruff <fletcherw@chromium.org>
Reviewed-by: Jesse Schettler <jschettler@chromium.org>
Reviewed-by: Jun Mukai <mukai@chromium.org>