Update patch set 3
Patch Set 3:

(4 comments)

The whole point of the present CL is that I have tried and failed to deliver on my promise. The test fails without testing what it is supposed to test. If you can find a way to make the test work, that is great. Otherwise, just know that I tried to keep my promise.

After the test shows the home launcher and then enters overview, the overview session appears empty even though the shelf shows that both ARC apps are still running. I am stumped because I cannot reproduce such behavior in manual testing on the same device. Here is a video of what I am talking about:
https://photos.app.goo.gl/WECaFiEUJ85KAc4J6

Patch-set: 3