BlockingSync: Split namespaces at \n
The for-loop which this CL modifies is intended to process one line at a
time from `namespaces`, a []byte of output from an earlier command.
But there was no processing to separate `namespaces` at newlines, so
instead the for-loop was processing one byte at a time... which is bad.
Needless to say, any code that hit this block would fail.
So, now we split at newlines before iterating.

BUG=b:155812675
TEST=firmware.BlockingSync on a sarien (or other NVMe device)

Change-Id: I057d9c87b34911e3c33c571bd5f41eaf18d7b373
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/2190980
Tested-by: Greg Edelston <gredelston@google.com>
Commit-Queue: Greg Edelston <gredelston@google.com>
Reviewed-by: Ricardo Quesada <ricardoq@chromium.org>