Update patch set 23
Patch Set 23:

(19 comments)

somewhere in the .go, or perhaps create a new file that tells from where the ds3.hid file was taken and the steps to update it in case it is needed.
also, what's the license of ds3.hid? the license must be honored.

Patch-set: 23
Reviewer: Gerrit User 1202079 <1202079@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>