Update patch set 1
Patch Set 1:

I thought about reducing the intervals in the tests again, but I am worried that the test may become flaky again. I think it should be a bit better with this CL, but it may not fix it entirely.

E.g., what if there happens to be high load on the machine in just the moment when the snapshot is requested and for some reason it is 200 ms delayed? Not sure how realistic this is though...

Patch-set: 1
Reviewer: Gerrit User 1375928 <1375928@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>