Reset the selected page at the end of ui.LauncherPageSwitchPerf
As described in crbug.com/1081285, the page of the apps-grid is
sticky in tablet-mode; it stays in the last selected page after
reopening. And with a chromium CL, this stickiness comes for
the tablet-transition. So when it selects a page and then
switches to the tablet-mode, the first shown apps-grid shows
the last selected page.

This comes with unexpected behavior when running both clamshell_mode
and tablet_mode of ui.LauncherPageSwitchPerf at the same time.

This CL fixes the problem by resetting the selected page at the
end of the test.

BUG=chromium:1081285
TEST=tast run ui.LauncherPageSwitchPerf'*' on eve

Change-Id: I501c416ac2db71d5b06699e3f8a64df1dd801096