[lacros Tast gpu tests] Create half-height window for moveocclusion_rot90.
This CL modifies the moveocclusion_rot90 test to create half-height
instead of half-width windows if the display is in a portrait
orientation. This fixes an issue where resizing a window on a portrait
display to be half-width would be smaller than the minimum window size
(which causes a test failure).

BUG=chromium:1053851
TEST=tast run <eve> lacros.GpuCUJ.*

Change-Id: If83d3355a1192886f37f820fd093bda4d71fa30e
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/2195235
Commit-Queue: Eliot Courtney <edcourtney@chromium.org>
Tested-by: Eliot Courtney <edcourtney@chromium.org>
Reviewed-by: Hidehiko Abe <hidehiko@chromium.org>