Run UIAutomator Test Using Tast For Netflix app
BUG=b:154835097
TEST=None

Tast Test for Netflix app verifies launch app
from the playstore, app is logged in and reaches the
main activity page of the app and sign out of an app.
Also for Clamshell Mode it verifies Clamshell Full Screen,
Minimize and Restore,Clamshell Resize, Reopen tests
and for Touchview Mode it verifies Minimize and Restore, Reopen tests.

Change-Id: Ide1bfc2fb0a58194702e2013a05eef05ef71182b