tast-tests: Add poll when extracting rootNode
When trying to extract the rootNode in a11yTree test, sometimes
the newly obtained tree contains the old root (ie. when activity is
EditTextActivity, MainActivity root is obtained), which is the cause of
flakiness. Adding a poll to give ample time to allow the tree to be
refreshed.

Ran a11yTree test, for a few hours and observed no failures.

BUG=b:156204447
TEST=tast run DUT_IP arc.AccessibilityTree

Change-Id: I3f06dfb617796e4efbef3ebc495624d00f352f50