wifi: Port network_WiFi_SimpleConnect.wifi_check{DFS,SSIDLimits,SGI}.
Port three sub-tests under network_WiFi_SimpleConnect to Tast.

Also introduce a utility for generating hostapd options with common
settings. It will be more helpful when there are more testcases that
don't quite care the protocol and channel. e.g. upcoming NonAsciiSSID
testcases. We can also apply the this to the security testcases in
SimpleConnect later if it is suitable.

BUG=chromium:1034875
TEST=`tast run $DUT wifi.SimpleConnect.*`

Change-Id: I806f45796c6561a14d3f56acfc71b5bf7b950a4a