wifi: Port more cases from network_WiFi_SimpleConnect
Port sub-tests under network_WiFi_SimpleConnect to Tast.
- wifi_checkDFS
- wifi_checkSSIDLimits
- wifi_checkSGI
- wifi_checkRawPMK

Also introduce a utility for generating hostapd options with common
settings. It will be more helpful when there are more testcases that
don't quite care the protocol and channel. e.g. upcoming NonAsciiSSID
testcases. We can also apply the function to the security testcases in
SimpleConnect later if it is suitable.

BUG=chromium:1034875
TEST=`tast run $DUT wifi.SimpleConnect.*`

Change-Id: I806f45796c6561a14d3f56acfc71b5bf7b950a4a