Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> I am using a mock timer now. Note, this CL is not working properly yet. There are still weird timing issues, where the test sometimes hangs forever.
> 
> Mocking the Sleep() seemed like a simple solution at first. But the complexity of the code has already increased a lot and it is hard to follow the code through goroutines, channels, fake Sleep() statements and context cancellation.
> 
> I think this code would be very hard to maintain and I am actually much more worried about the correctness of the test than of the correctness of the actual code.
> 
> Maybe this is a bit overkill? What we are testing here is a unit test for the testing framework. I think we should reconsider simply increasing the Sleep time durations. Or if that's not acceptable, maybe deleting this test entirely. What do you think?

I don't think it is acceptable to increase timeouts... that is the worst possible way to resolve test flakiness.

I definitely agree that it is generally difficult to test the behavior of Timeline. One possible way would be to give up testing Timeline's behavior about taking snapshots at a regular interval; but I believe we still want to test that Setup and Start are called once and Snapshot is called at least once.

I still believe that it's possible to test the regular interval behavior with a fake clock. We have to carefully design a fake clock though. It's a good exercise for concurrent programming.

Patch-set: 2
Reviewer: Gerrit User 1118346 <1118346@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>