Set timeout to ash.WaitForOverviewState
ash.WaitForOverviewState does not have the timeout. When the system
UI comes to an unexpected state, it will wait for the expiry of
the given context. The wait will be as long as multiple minutes.

This CL provides the timeout and gives up when the given timeout
has passed. This helps for tests to fail timely and reduce the
total time of running tests.

BUG=chromium:1081974
TEST=tast ui.HotseatAnimation'*' ui.TaskSwitchCUJ.tablet_mode

Change-Id: Ia617f8e3cc68d99e11c28ee426351d1ea9aab229