Tast: rename WifiCheckSarTable to WifiCheckIntelSarTable
This test only works when run on a device which uses Intel WiFi.
Previously, it was thought that this test may eventually be extended to
support other WiFi vendors, but now the plan is to implement tests for
other vendors elsewhere. For the sake of being explicit (and reducing
confusion points with partners) rename this test to accurately convey
that it is only intended for use with Intel WiFi.

BUG=None
TEST=`tast run $DUT network.WifiCheckIntelSARTable`

Change-Id: I8c3762e35344c6a2400185ded180b563e5fb964c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/2200122
Reviewed-by: Kevin Lund <kglund@google.com>
Reviewed-by: Shuo-Peng Liao <deanliao@chromium.org>
Reviewed-by: Keigo Oka <oka@chromium.org>
Reviewed-by: Ricardo Quesada <ricardoq@chromium.org>
Reviewed-by: Brian Norris <briannorris@chromium.org>
Tested-by: Kevin Lund <kglund@google.com>
Commit-Queue: Kevin Lund <kglund@google.com>