Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> > Patch Set 2:
> > 
> > to promote it to CQ, the test must have passed the last 20 builds:
> > https://chrome-internal.googlesource.com/chromeos/chromeos-admin/+/master/doc/tast_add_test.md#step-4_remove-attribute-to-make-it-run-on-infra
> > 
> > it seems it is not the case for arc.Optin: https://tastboard.corp.google.com/test/arc.Optin
> 
> This test is based on already existing functionality. It is just simplified version of https://tastboard.corp.google.com/test/arc.PlayStore (and others). I think it is bit different from other tests, developed from scratch. I understand you follow the defined rules, but this one is bit different. Anyway, leaving this up to you.

It seems that arc.PlayStore hasn't passed the last 20 builds too.
Yes, I'm just following the "playbook" here (must pass 20 builds in a row).

The risk of promoting a test to CQ when it is not ready, is that it will block the builds.

arc.Optin has been stable for the last 6 builds, but before that it was kind of flaky. So, let it run for a couple more days. Probably by Monday/Tuesday we already have 20 builds and I'll +2.

Patch-set: 2
Reviewer: Gerrit User 1202079 <1202079@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>