tast-tests: Use map to store RAPL values instead of struct members
This CL cleans the implementation of rapl.go using map instead of
struct since it is hard to iterate over the values stored in struct
on reporting metrics and RAPL values are vary on different CPUs.
This CL also introduces Total() and Uncore() for RAPLValues to be
used in gpu_cuj.go.

BUG=b:153028627
TEST=RAPL metrics report the same values as before this CL

Change-Id: I1df219b593896a334261858f08644f988afcd08b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/2206559
Reviewed-by: Ricardo Quesada <ricardoq@chromium.org>
Tested-by: Hikaru Nishida <hikalium@chromium.org>
Commit-Queue: Hikaru Nishida <hikalium@chromium.org>