Don't fail UncleanShutdown on old kernel crashes.
If the boot collector detects a kernel crash, it will create a
'kernel-crash-detected' file instead of an 'unclean-shutdown-detected'
file, so the unclean shutdown isn't counted.

Sometimes, DUTs happen to have recoverable BIOS panics on the boot
before the crash test runs, for reasons out of the telemetry team's
control. These cases were causing platform.UncleanShutdownCollector
to flake. To work around these flakes, manually touch the
unclean-shutdown-detected file if the kernel-crash-detected file is
present.

Ideally, we should log a warning in this case (otherwise we won't catch
a bug where the boot collector *always* creates kernel-crash-detected
instead of unclean-shutdown-detected), but that's currently not possible
in tast; see crbug.com/1083533 for a feature request.

BUG=chromium:1083533,chromium:1041639
TEST=tast run $DUT platform.UncleanShutdownCollector*

Change-Id: Idda2f39b0bb40c24c56eddebc040634dfaca41c0