wificell: hostapd: Comment out Greenfield option.
With the discussion in crrev.com/c/1989649 and crrev.com/c/2200521,
as our test APs don't support GF capability, we should avoid the usage
of it. Commenting out the enum with reasons might be better than
ignoring it or failing in runtime as the user can know it without
running.

BUG=None
TEST=fast_build.sh -T

Change-Id: I95b31d8fdbac0958579ea1c577d08660dd4b26e7
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/2206564
Tested-by: Yen-lin Lai <yenlinlai@google.com>
Commit-Queue: Yen-lin Lai <yenlinlai@google.com>
Reviewed-by: Matthew Wang <matthewmwang@chromium.org>
Reviewed-by: Hidehiko Abe <hidehiko@chromium.org>