wificell: hostapd: Comment out Greenfield option.
With the discussion in crrev.com/c/1989649 and crrev.com/c/2200521,
as our test APs don't support GF capability, we should avoid the usage
of it. Invalidate the enum with reasons in comments might be better
than ignoring it or failing in runtime as the user can know it without
running.

BUG=None
TEST=fast_build.sh -T

Change-Id: I95b31d8fdbac0958579ea1c577d08660dd4b26e7