test-tests: Use well known path for Drive smoke test
Currently, Drive smoke tests create a temporary file and checks it
appears in the UI. However, sometimes this temporary file gets synced to
the Drive server, but its deletion does not have time to be synced as it
happens at the end of the test. This causes temporary files to be left
behind on Drive. Use a well known path for this file instead of a
temporary file so it does not matter if the file gets left behind.

BUG=chromium:1082662
TEST=tast run -build $DUT platform.Drivefs

Change-Id: I45b42fe19e4d424b508fa58ed79293c8eb5be7ca
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/2206567
Reviewed-by: Sergei Datsenko <dats@chromium.org>
Reviewed-by: Keigo Oka <oka@chromium.org>
Tested-by: Austin Tankiang <austinct@chromium.org>
Commit-Queue: Austin Tankiang <austinct@chromium.org>