tast-tests/video: merge PlaybackVDPerf into PlaybackPerf
This CL merges PlaybackVDPerf.bla into PlaybackPerf as
bla_alt cases, following the taxonomy of video.Play and
video.Seek. This is needed for further CLs down the line,
but it also makes sense in itself because the test
cases are the same with a different test param, so they
don't merit a file on its own.

No new functionality intended, only code refactoring.

BUG=chromium:1071841
TEST=tast run DUT video.PlaybackPerf.*_alt on kohaku

Change-Id: I5ea39d300253c10f7a8e72f410521c63c56380c8
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/2209442
Tested-by: Miguel Casas <mcasas@chromium.org>
Reviewed-by: Brendan Hansknecht <bhansknecht@chromium.org>
Commit-Queue: Miguel Casas <mcasas@chromium.org>