Speculative fix of WaitForLocationChangeCompleted
Some tests may fail on WaitForLocationChangeCompleted, claiming
there are still some of location change events. Thus this increases
the entire timeout. Most of the cases, this won't add any time
to the callers since it exists if there are no events on the
regular timeout (2 seconds). entireTimeout is the total cap of
the function.

BUG=chromium:1084744
TEST=tast run ui.LauncherDragPerf

Change-Id: I98d33907425f424e52cd1a98dee9dca4cc2fca7d