Speculative fix of LauncherPageSwitchPerf on samus, guado
Sometimes ui.LauncherPageSwitchPerf fails on clicking the page
switch button. One possible hypothesis which matches with the
screenshot is, it clicks on a wrong location so it does not cause
page switch at all.

This can happen when the location cache of chrome/ui package is
not updated timely. This CL addresses this by adding
WaitForLocationChangesToComplete.

BUG=chromium:1084153
TEST=tast run ui.LauncherPageSwitchPerf'*'

Change-Id: Iead1fab5a8b47bd4f61f854fdbefc10383bbe9c8
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/2210939
Tested-by: Jun Mukai <mukai@chromium.org>
Reviewed-by: Andrew Xu <andrewxu@chromium.org>
Reviewed-by: Katherine Threlkeld <kathrelkeld@chromium.org>
Commit-Queue: Jun Mukai <mukai@chromium.org>