Speculative fix of LauncherPageSwitchPerf on samus, guado
Sometimes ui.LauncherPageSwitchPerf fails on clicking the page
switch button. One possible hypothesis which matches with the
screenshot is, it clicks on a wrong location so it does not cause
page switch at all.

This can happen when the location cache of chrome/ui package is
not updated timely. This CL addresses this by adding
WaitForLocationChangesToComplete.

BUG=chromium:1084153
TEST=tast run ui.LauncherPageSwitchPerf'*'

Change-Id: Iead1fab5a8b47bd4f61f854fdbefc10383bbe9c8