tast-tests: Fix ui.SplitViewResizePerf.clamshell_mode on chromebox
Here is a speculative fix for ui.SplitViewResizePerf.clamshell_mode on a
chromebox with an external touch screen. The test shall get info for the
primary display instead of the internal display. For consistency and
simplicity, ui.SplitViewResizePerf shall do the same although it does
not run on the chromebox (the chromebox does not support tablet mode).

BUG=chromium:1084784
TEST=ui.SplitViewResizePerf,ui.SplitViewResizePerf.clamshell_mode

Change-Id: I90153ab6f6e6a1ee88a4d3ba27ee6322e9035653
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/2212672
Reviewed-by: Jun Mukai <mukai@chromium.org>
Tested-by: Avery Musbach <amusbach@chromium.org>
Commit-Queue: Avery Musbach <amusbach@chromium.org>