Fix Files app OpenDownloads function
The OpenDownloads function, used to navigate to the Downloads
folder in tast tests was selecting "Downloads" in the directory
tree. If Downloads had nested folders, running this function would
select  the entire "Downloads" subtree and select the middle row of
this subtree, which often resulted in navigating in a subdirectory
of Downloads.

This change fixes this by clicking on the Downloads row in the
Downloads subtree.

BUG=chromium:407010
TEST=tast run DUT_IP ui.FilesAppImageQuickView
TEST=tast run DUT_IP platform.DriveFs

Change-Id: If509cdfc49d0f07ba0220f41b796b48b3698869a