Update patch set 10
Patch Set 10:

(9 comments)

Thanks a lot for your review.
I will look into creating a CL in between to move the previous filesapp tests to a new package, probably "files". I think that importing the package "filesapp" in this test, and make this test part of this same package wouldn't make sense for the compiler.

Patch-set: 10