Update patch set 3
Patch Set 3:

Some context: There were two tests Suggestions and SuggestionsInformational. Suggestions tested English and SuggestionsInformational tested CJK. English is special because when the decoder is not ready, any key presses will be dropped. So we have to keep tapping 'a' until the decoder is ready. For CJK, we would queue up key presses so it's not a problem.

IIUC your change would only test CJK and not English anymore right? I would prefer if we can have test coverage for English as well because it is by far our biggest usage.

Patch-set: 3
CC: Gerrit User 1197444 <1197444@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>