tast-tests: clarify Video preconditions
Chrome has two hardware accelerated video decoder implementations:
a "legacy" (in use) and a "new" (upcoming) VideoDecoder. The
"new" one is being launched in a few devices (see e.g. [1] and
crrev.com/c/2214764).

ToT Chrome Video preconditions force-enable and -disable the "new"
video decoder, but post launch we'd like to change this to having
preconditions to either let the "main" video decoder be used or to
force the "alternate" implementation. Most of the test cases will
exercise the "main" (i.e. what the users see) and we'll keep a few
with the "alternate" implementation.

[1] https://pastebin.com/C0VsunB6

BUG=chromium:1050248
TEST=tast run ${IP} video.Play.*vp* video.Seek.*vp* on kohaku

Change-Id: I3931d01acdd00c5564f3688dddaae7e079e40bf7