tast-tests: clarify old/new Video decoder preconditions
Chrome has two hardware accelerated video decoder implementations:
a "legacy" (in use) and a "new" (upcoming) VideoDecoder. The "new"
one is being launched (crrev.com/c/2214764) in a few devices
(non trivial ones: pastebin.com/Ext3CJHy).

ToT Chrome Video preconditions force-enable and -disable the "new"
video decoder, but post launch we'd like to change this to having
preconditions to either let the "main" video decoder be used or to
force the "alternate" implementation. Most of the test cases will
exercise the "main" (i.e. what the users see) and we'll keep a few
with the "alternate" implementation.

This CL refactors the ChromeVideo and ChromeVideoVD into such
"main" and "alternate" schemes, depending on the presence or
absence of the USE flag. More discussion about this pattern can
be found in
https://groups.google.com/a/google.com/g/chromeos-gfx-video/c/asO-BR0b4-U/m/7hQYuoUlGgAJ

BUG=chromium:1050248
TEST=tast run ${IP} video.Play.*vp* video.Seek.*vp* on kohaku

Change-Id: I3931d01acdd00c5564f3688dddaae7e079e40bf7
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/2216300
Tested-by: Miguel Casas <mcasas@chromium.org>
Reviewed-by: Jun Mukai <mukai@chromium.org>
Reviewed-by: Po-Hsien Wang <pwang@chromium.org>
Reviewed-by: Ilja H. Friedel <ihf@chromium.org>
Commit-Queue: Miguel Casas <mcasas@chromium.org>