tast-tests: Rework EvalPromise in arc.SettingsBridge.
This CL introduces chromeproxy package, which is go binding of
chrome.proxy API.
In the clean up, importantly, we get rid of EvalPromise, which is
already deprecated. In addition, we get rid of script construction
by fmt.Sprintf.

In addition, this CL intorduces s.Run() for table driven test
to make error log more developer familiar.

BUG=chromium:2215770
TEST=Ran arc.SettingsBridge locally.

Change-Id: I99a47b4c94ec0ffd30cc55c86c29c093208e1b6f