tast-tests: Add DragMaximizedWindowPerf test.
Tests dragging to unmaximize, which does shrinks the window with an
animation while updating bounds and unoccluding other windows. We may
want to use some more specific histograms in the future, but support
to add those does not exist in Chrome yet, so use the default cross
fade animation histogram for now.

BUG=chromium:1081986
TEST=tast -vebose run <IP> ui.DragMaximizedWindowPerf

Change-Id: Iaf36161e51d14a2db1b47505c33064b2dd807e10