tast: Create package for nebraska + dlc
Nebraska is a generic server that can be used for both
updates/installations. Seperating Nebraska logic in dlcservice.go to be
leveraged in future tests as well as accepting context.Context and
testing.State as arguments when using subtests.

DLC package is used for communicating with dlcservice_util (the
executable that is a commandline interface with dlcservice daemon).

Also each subtest is scoped better to its own testing.State and
context.Context when run from the main test.

BUG=none
TEST=tast run $IP platform.DLCService

Change-Id: Id792d5e75eb417e4577bf58c2cee828f162b736b