tast-tests/webrtc: refactor MeasureGPUCounters out its callsites
This CL moves the code of MeasureGPUCounters out of its two
current callsites in preparation of using it in video/ perf
tests. To simplify things, the code is moved out of dri.go
(since it doesn't actually interact with the DRI API) and into
a new file perf.go.

No new functionality intended, only code refactoring.

BUG=chromium:1071841
TEST=tast run DUT webrtc.*Perf.vp9* on kohaku

Change-Id: I844869d0bccf63adcee6f0dfb1ed03666b87d369