Update patch set 1
Patch Set 1:

> This test is quite reliable, but not perfect (see http://tastboard/test/printer.PrintUSB).

Based on the other tests I see from mainline tests on Tastboard, I think that this is within the acceptable range, so CR+1.

> I've looked into the "Test did not finish" errors, but it's not clear what they mean. Is it equivalent to a context deadline being exceeded?

That message seems to be raised by autotest/server/site_tests/tast/tast.py, when the Tast test returns with an end_time of 0. I'm not sure why that would happen, nor whether that is equivalent to a context deadline being exceeded.

> If so, it would suggest that the last operation in the log is hanging.

I think it is not the last operation that is hanging, because the logs don't show . For example, here are the logs from the R85-13156.0.0 eve failure:

https://00e9e64bac779112d3a7e70b59cb8fc059dae6831d47e530aa-apidata.googleusercontent.com/download/storage/v1/b/chromeos-autotest-results/o/swarming-4c5c72f00f463f10%2Fautoserv_test%2Ftast%2Fresults%2Ftests%2Fprinter.PrintUSB%2Flog.txt?qk=AD5uMEuZqkUbmIawqKiZ7mqFkMsRIqzEiQFODkeRHJt828k29pI3ntPETqhNlVko7XKt_9bt0hca71xgBNxmEdPp8LUwuhAgmgmS4Vs5FbgulMgYt6CqXoZGWpSPBJpRJhVUwtH75g-3KPFB9N0Q3SQT3xl6PbR-CCpntlTz2lRuFoUHvHSZxkJBqUYEwcYyOwZeuyHftp108mPjCE5bK3m5KJq3znPyM2rfpvnOvhBNffRsNUzwlQFHpzQCM-5vjNEKcCbk4uC8nXLJY2kReiN5e3XCw8lcLeUF3mTNamSAaXOyDVZocoozFBG6M8sZTmTX7GWJgwbL1Vdd2jpIIcWlWJsNNmb_KXZkqbnaNtW73sPLvFX3lZiazKn0cowZJbSkTrX_W14crleTJveCDYGiz83SA-O5D_mGka1ykckyliG_L8_x2gTjUzt0UZ3ybEhUN6yRFvvz8X3AS6d_oxP-I3XblN764BlPE_gddc5pcv7rP0d9YsocFLPd7e6QiZHFZHVZu2rySV_i5oFUzxC-N0Ov1C6sND8Ny44sRT4s_AGJ27qhpWVKSoEGouz1l8yDECsoKViDaGyjeOvKvy5XSTH_Wrhaja3yvQdiwd4VILKVFXmfN56nT0eqPTuwoJ6Ke8_diGmCJcYF1Ve-1ewiQR4shfT_w_n2ds1xEysTs3qFEkzRtqNHevWmbZO7WRT1H1gVVjUziCFEIfR-doCtffw7G5NkSSIE0q5CMU1neyuwI5tzsZ8lU79_K7qUNNjCkBQRm-b3y3XGTmKfeFt4R5omkAaC8Eiz_CJW8c2YEGl2Lyag_39yN-Hh2oQH00uZ9qpqK-gQC8HLqvDWBiy-sJa_cPnPrtLQNAmcGnW28tP1nzPAFTaJjlbRR5ACVwQQZLyM4VSkVDchpOAPCktmID6jIgxORvwFCe2nCYboBW657x3jnTg&isca=1

That said, the

Patch-set: 1
Reviewer: Gerrit User 1326157 <1326157@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>