tast-tests: snapshot: implement SnapshotManager
This CL is useful to preserve the important data that gets wiped in
production but necessary for integration test. For example, tpm owner
password.

In order to test this file, the child CLs are necessary. Th validation
is done by observing if the file content, permission, and the UID/GUID
is successfully restored.

BUG=b:154672602
TEST=see above.

Change-Id: I912180b5a5d75a685868352370fcc52df70a4a89