tast-tests: ready: system-level Snapshot manager stores tpm data
This CL preserves the tpm owner password and is useful to rewind hwsec
userland daemons to a state at early stage.

In order to test this file, the child CLs are necessary. Th validation
is done by observing if the wiped owner password can actually come back
by restoring the snapshot at the beginning of a test. In this case, we
choose hwsec.AttestationNoExternalServer to see if the enrollment status
gets wiped and performs enrollment again. Also, we clear the owner
password between 2 tast runs and confirm that the second run can
recover the tpm local data from the backup by the first run.

BUG=b:154672602
TEST=see above.

Change-Id: Ie81521a8bfe80680c0aeab7bd91c7ddffd7af24b