tast-tests: hwsec: also back up secrets in case of soft clear
This will used to back up tpm manager local data in ready.go and used to
rewind the state machine of tpm manager whenever necessary.

In order to test this CL, we run a test item with soft-clear, and then
run hwsec.AttestationNoExternalServer, so we can verify that the newly
generated secrets after soft-clear is also able to be recovered from.

BUG=b:154672602
TEST=see above.

Change-Id: Ibd33a0b33d28ff6858d1aa4d57492532d4f5c4e6
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/2237568
Tested-by: Leo Lai <cylai@google.com>
Commit-Queue: Leo Lai <cylai@google.com>
Reviewed-by: Shuhei Takahashi <nya@chromium.org>