tast-tests: hwsec: tast-side fake pca agent control
The previous solution tries to add a new upstart conf but the
installation, even if it works, contaminated /etc/init. Since it is just
for testing purpose, it's better to avoid that.

Thus, implementing a simple fake pca agent service control at tast side
is the way to go.

BUG=b:154672602
TEST=hwsec.AttestationNoExternalServer

Change-Id: I75cf98ed0df3e1f1b48542dbd99ec892abdf9b90
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/2220931
Tested-by: Leo Lai <cylai@google.com>
Commit-Queue: Leo Lai <cylai@google.com>
Reviewed-by: Shuhei Takahashi <nya@chromium.org>