tast-tests: hwsec: attestation local infra enabled after readiness
It makes more sense to check TPM readiness first because normally
ready.go should take TPM ownership; if there is no ownership, there is
no way the local infra can be enabled. It also makes sense when
considering TPM readiness checking it more light-weighted, so it should
be the precondition.

BUG=b:154672602
TEST=hwsec.AttestationNoExternalServer

Change-Id: I0c7d92160e4679172cc36346505ef9a7d1ce9ada