tast-tests: hwsec: functions that backup/restore tpm manager data.
This will used to back up tpm manager local data in ready.go and used to
rewind the state machine of tpm manager whenever necessary.

In order to test this file, the child CLs are necessary. Th validation
is done by observing if the wiped owner password can actually come back
by restoring the snapshot at the beginning of a test. In this case, we
choose hwsec.AttestationNoExternalServer to see if the enrollment status
gets wiped and performs enrollment again. Also, we clear the owner
password between 2 tast runs and confirm that the second run can
recover the tpm local data from the backup by the first run.

BUG=b:154672602
TEST=see above.

Change-Id: Ie4406ea577bbbd726f2930f57118d64c3f22864a