tast-tests: Use local installer instead of artifact for factory.Goofy
We have added the factory toolkit installer to the test image.
We are revisiting the design of factory_image.zip and may remove it in
artifacts. Also, this change will decouple this Tast test and the
dependency issue in chromite which make this test so flaky currently.

BUG=b:150189948
TEST=passed preCQ
TEST=tast run $DUT factory.Goofy

Change-Id: I9e37fa5d505bfd6210afa0bb1664eafa63407071
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/2217661
Reviewed-by: Shuhei Takahashi <nya@chromium.org>
Commit-Queue: Meng-Huan Yu <menghuan@chromium.org>
Tested-by: Meng-Huan Yu <menghuan@chromium.org>