Update patch set 2
Patch Set 2:

As discussed in chat with Miriam:

Policy tests run in 15 seconds (completely) when I test it. The timeout also only applies to the precondition which runs before the test.
Since no other test is hitting this timeout I don't think we should increase it.
If we ever hit that timeout the better solution will be to make login faster 🙂. Waiting for 1 minute for login is not a great user experience.

Patch-set: 2