power: Add tests to labqual group
This commit adds the following tests to labqual:
  power.CheckStatus
  power.Reboot

A subset of tests is used for qualifying a model before deploying it to
the lab. Previously, a wide group of tast tests were run, but that group
is being narrowed to prevent unnecessary delay. This commit ensures
that necessary tests are performed for lab qualification testing.

BUG=b:155173602
TEST=None

Change-Id: I79bf6c51c32bfdae6b23876425f2cf57164a35dd