platform: Add tests to group:labqual
This commit adds the following tests to the labqual group:
  platform.CrosConfig
  platform.Crossystem
  platform.TPMResponsive

A subset of tests is used for qualifying a model before deploying it to
the lab. Previously, a wide group of tast tests were run, but that group
is being narrowed to prevent unnecessary delay. This commit ensures
that necessary tests are performed for lab qualification testing.

BUG=b:155173602
TEST=None

Change-Id: I8ad6aed3160b6ab529df5ec4d78e76da9b555b9b