platform: Add tests to group:labqual
This commit adds the following tests to the labqual group:
  platform.CrosConfig
  platform.Crossystem
  platform.TPMResponsive

A subset of tests is used for qualifying a model before deploying it to
the lab. Previously, a wide group of tast tests were run, but that group
is being narrowed to prevent unnecessary delay. This commit ensures
that necessary tests are performed for lab qualification testing.

BUG=b:155173602
TEST=tast -verbose run $DUT '(!disabled && "group:labqual")'

Cq-Depend: chromium:2220333
Change-Id: I8ad6aed3160b6ab529df5ec4d78e76da9b555b9b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/2220287
Tested-by: Justin TerAvest <teravest@chromium.org>
Reviewed-by: Rob Barnes <robbarnes@google.com>
Reviewed-by: Oleh Lamzin <lamzin@google.com>
Reviewed-by: Jack Rosenthal <jrosenth@chromium.org>