firmware: add tests to group:labqual
This commit adds the following tests to the labqual group:
  firmware.Flashrom
  firmware.GSCtool

A subset of tests is used for qualifying a model before deploying it to
the lab. Previously, a wide group of tast tests were run, but that group
is being narrowed to prevent unnecessary delay. This commit ensures
that necessary tests are performed for lab qualification testing.

BUG=b:155173602
TEST=None

Cq-Depend: chromium:2220333
Change-Id: Ia2efe4c8396cbfa97e08b64986ce3142d7c4cf6e