hwsec: add tests to group:labqual
This commit adds the following tests to the labqual group:
  hwsec.Login
  hwsec.LoginGuest

A subset of tests is used for qualifying a model before deploying it to
the lab. Previously, a wide group of tast tests were run, but that group
is being narrowed to prevent unnecessary delay. This commit ensures
that necessary tests are performed for lab qualification testing.

BUG=b:155173602
TEST=None

Cq-Depend: chromium:2220333
Change-Id: Ibc85751403b07fbedbf27db5b00a62fa0d99d0b9