hwsec: add tests to group:labqual
This commit adds the following tests to the labqual group:
  hwsec.Login
  hwsec.LoginGuest

A subset of tests is used for qualifying a model before deploying it to
the lab. Previously, a wide group of tast tests were run, but that group
is being narrowed to prevent unnecessary delay. This commit ensures
that necessary tests are performed for lab qualification testing.

BUG=b:155173602
TEST=tast -verbose run $DUT '(!disabled && "group:labqual")'

Cq-Depend: chromium:2220333
Change-Id: Ibc85751403b07fbedbf27db5b00a62fa0d99d0b9
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/2220289
Reviewed-by: Greg Edelston <gredelston@google.com>
Reviewed-by: Garry Wang <xianuowang@chromium.org>
Reviewed-by: Hidehiko Abe <hidehiko@chromium.org>
Reviewed-by: Rob Barnes <robbarnes@google.com>
Tested-by: Rob Barnes <robbarnes@google.com>