Update patch set 1
Patch Set 1: Commit-Queue+1

> Patch Set 1: Code-Review-1
> 
> (1 comment)
> 
> Looks like this test is not stable enough. Message from Tast board:
> 
> This informational test failed very recently. Every mainline test should eventually be promoted to the Commit Queue so it can prevent breakages. Please stabilize the test as soon as possible (go/tast-add-test).
> 
> 
> So, let's wait until test will be stable enough: 20 success builds in a row.

I'm not sure that waiting 10 days (we seem to be averaging two builds a day) is worth it. The only failures since the beginning of May are the one fixed by the CL mentioned in the description, and the flakiness tracked by crbug.com/949387, which exceeds this test and affects any test that logs in or mounts cryptohomes.

https://docs.google.com/document/d/1TTA62izAYHylqPSFc9hyNeqr0IEVz200C3Nzs05aEK4/edit is not practical if we need to wait 10 days to move tests out of informational. Given that the CL fixing things landed over the weekend, my preference would be to land this before end-of-week. We know that the CL fixes the flake, because it addresses it explicitly.

Patch-set: 1
Reviewer: Gerrit User 1000985 <1000985@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Commit-Queue=+1