Update patch set 8
Patch Set 8:

(2 comments)

I know you're just porting the test over from Autotest, but I feel like I may as well bring up any issues I have with this test while it has my attention. I don't really like that this test is just sending a CSA off and checking for disconnects, as opposed to checking that we disconnect because we received a CSA. If we somehow disconnect for some unknown/unrelated reason, we still pass it here. In fact, doing this 10 times increases the likelihood that this happens.

How much work would it be to rewrite this test to be less hacky? If it requires too much extra infrastructure/framework, we can just leave a TODO and spawn a bug.

Patch-set: 8
Reviewer: Gerrit User 1243287 <1243287@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>