Move some crash tests from platform/to crash/.
This CL doesn't move UncleanShutdownCollector or EphemeralCrashCollector
because we're waiting on 20 successful runs for them, and it's easier to
count those if we don't move them in the middle of waiting for the
successful runs.  It also doesn't move any platform.UserCrash.* tests
because they depend on a library in the platform/ sub-directory and so
moving them will be (slightly) more involved.

This CL renames UdevCrash to Udev to avoid redundancy (crash.UdevCrash),
but leaves crash.ReporterCrash with the Crash at the end to make clear
that it isn't testing some other property of the crash reporter.

BUG=chromium:1038744
TEST=tast run $DUT crash.{ReporterCrash,KernelWarning,\
  SelinuxViolation,ServiceFailure,Udev,SuspendFailure,DevCoredump}*

Change-Id: Ia3473de644de94650550c60b741cdd6635290afa